🚚 Moved other setting's vue files in HOME
