+++ src/components/World/HOME/Others/BootSettingsView.vue
