+++ src/pages/WorldTabs/HomePage.vue
@@ -2,8 +2,8 @@
 import { ref } from 'vue';
 import { setScrollTop } from 'src/components/World/HOME/scroll';
 import WorldIconView from 'src/components/World/HOME/Top/WorldIconView.vue';
-import BootSettingsView from 'src/components/World/HOME/OtherSettings/BootSettingsView.vue';
-import WorldDeleteView from 'src/components/World/HOME/WorldOperation/WorldDeleteView.vue';
+import BootSettingsView from 'src/components/World/HOME/Others/BootSettingsView.vue';
+import WorldDeleteView from 'src/components/World/HOME/Others/WorldDeleteView.vue';
 import WorldNameView from 'src/components/World/HOME/Top/WorldNameView.vue';
 import VersionSelecterView from 'src/components/World/HOME/Top/VersionSelecterView.vue';
 import RunningBtn from 'src/components/World/HOME/RunningBtn.vue';
@@ -28,6 +28,7 @@ setScrollTop(scrollTop)
     <div class="mainField">
       <RunningBtn to="/console" :test-font-size="1.1" class="full-width q-mt-xl" />
 
+      <!-- TOP -->
       <div class="row justify-center q-mt-sm q-pb-lg q-gutter-lg">
         <div class="col" style="min-width: 12rem;">
           <h1 class="q-pt-none">{{ $t("home.worldName.title") }}</h1>
@@ -38,7 +39,8 @@ setScrollTop(scrollTop)
         </div>
         <WorldIconView />
       </div>
-        
+      
+      <!-- Others -->
       <h1 class="q-pt-lg">{{ $t('home.setting.title') }}</h1>
       <BootSettingsView />
